move jwt to separate package and minor cleanup
